knockknock

+++ src/main/java/entities/Location.java
@@ -5,8 +5,8 @@ import java.util.Objects;
 public class Location {
     private int id;
     private String address;
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
 
     public Location(int id, String address, float latitude, float longitude) {
         this.id = id;
@@ -38,7 +38,7 @@ public class Location {
         this.address = address;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
@@ -46,7 +46,7 @@ public class Location {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
@@ -58,7 +58,7 @@ public class Location {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof Location location)) return false;
-        return getId() == location.getId() && Double.compare(getLatitude(), location.getLatitude()) == 0 && Double.compare(getLongitude(), location.getLongitude()) == 0 && Objects.equals(getAddress(), location.getAddress());
+        return getId() == location.getId() && Float.compare(getLatitude(), location.getLatitude()) == 0 && Float.compare(getLongitude(), location.getLongitude()) == 0 && Objects.equals(getAddress(), location.getAddress());
     }
 
     @Override
